CostsAndIncomes-component-nesting-logic-setting
