+++ src/components/CostsAndIncomesBox/CostsAndIncomesButtons/index.js
@@ -1,0 +1,1 @@
+export {default} from './CostsAndIncomesButtons'
