+++ src/components/CostsAndIncomesBox/_CostsAndIncomesBox.module.scss
@@ -1,0 +1,12 @@
+
+.content-box {
+    margin-left: auto;
+    margin-right: auto;
+    max-width: 1098px;
+    background-color: white;
+    padding: 32px 32px 62px 32px;
+    box-shadow: 0px 10px 60px rgba(170, 178, 197, 0.2);
+border-radius: 0px 30px 30px 30px;
+}
+
+
