+++ src/pages/CostsAndIncomesPage/CostsAndIncomesPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import CostsAndIncomesBox from '../../components/CostsAndIncomesBox/CostsAndIncomesBox'
+
+const CostsAndIncomesPage = () => {
+  return (
+    <section className={'page'}>
+        <CostsAndIncomesBox />
+    </section>
+
+  );
+};
+
+export default CostsAndIncomesPage;
